[Feature] Rename utils to helpers
